services/auth-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
python-jose[cryptography]==3.5.0
pydantic-settings==2.10.1
passlib[bcrypt]==1.7.4
email-validator==2.2.0
